requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Faker>=2.0,<=2.0.4
djangorestframework==3.9.3
